카드 기능 수정/추가
카드 이동 제외한 모든 기능 구현, 작동 확인 완료
조건/예외처리 부족

+++ src/main/java/com/sparta/developmentgroup1/comment/service/CommentService.java
@@ -6,6 +6,7 @@ import com.sparta.developmentgroup1.comment.dto.CommentRequestDto;
 import com.sparta.developmentgroup1.comment.dto.CommentResponseDto;
 import com.sparta.developmentgroup1.comment.entity.Comment;
 import com.sparta.developmentgroup1.comment.repository.CommentRepository;
+import com.sparta.developmentgroup1.user.entity.User;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -17,30 +18,35 @@ import org.springframework.transaction.annotation.Transactional;
 public class CommentService {
     private final CardRepository cardRepository;
     private final CommentRepository commentRepository;
-    // 댓글 생성
-    public CommentResponseDto createComment(Long cardId, CommentRequestDto requestDto) {
-        Card card = findCard(cardId);
-        Comment comment = new Comment(requestDto, card);
 
-        comment.setCard(card); //조회한 카드 받고
-        comment.setContent(comment.getContent());
-        comment.setCreator(comment.getCreator());
-         var savedCardComment = commentRepository.save(comment); //DB에 저장
-        return new CommentResponseDto(savedCardComment); //CardCommentResponseDto 생성자를 통해 필드 추가
+    // 댓글 생성
+    public CommentResponseDto createComment(Long cardId, CommentRequestDto requestDto, User user) {
+        Card card = findCard(cardId);
+        Comment comment = new Comment(requestDto, card, user);
+
+         commentRepository.save(comment); //DB에 저장
+        return new CommentResponseDto(comment); //CardCommentResponseDto 생성자를 통해 필드 추가
     }
+
     @Transactional //댓글 수정
-    public void updateComment(Long id, CommentResponseDto requestDto) {
+    public void updateComment(Long id, CommentRequestDto requestDto, User user) {
         Comment comment = findComment(id); // 댓글 찾기
 
-        comment.setContent(requestDto.getContent());
+        if (!comment.getCreator().equals(user)) {
+            throw new IllegalArgumentException("작성자 본인만 수정할 수 있습니다.");
+        }
+
+        comment.updateContent(requestDto.getContent());
 
         commentRepository.save(comment); // 변경된 카드 저장
     }
     @Transactional
-    public void deleteComment(Long id, CommentResponseDto requestDto) {
+    public void deleteComment(Long id, User user) {
         Comment comment = findComment(id); // 댓글 찾기
 
-        comment.setContent(requestDto.getContent());
+        if (!comment.getCreator().equals(user)) {
+            throw new IllegalArgumentException("작성자 본인만 수정할 수 있습니다.");
+        }
 
         commentRepository.delete(comment); // 변경된 카드 저장
     }
